// src/modules/email/enums/email-filter-key.enum.ts
export enum EmailFilterKey {
  SEEN = 'seen',
  NEW = 'new',
  SUBJECT = 'subject',
  SINCE = 'since',
  BEFORE = 'before',
  FROM = 'from',
  TO = 'to',
  CC = 'cc',
  BCC = 'bcc',
  ANSWERED = 'answered',
  DELETED = 'deleted',
  DRAFT = 'draft',
  WITH_FLAG = 'withFlag',
  WITHOUT_FLAG = 'withoutFlag',
  THREAD_ID = 'threadId'
}